prompt refactoring / work around bash 4.0 PROMPT_COMMAND issue
There's what appears to be a strange bug in bash 4.0 where using the
"$()" interpolation syntax from within a function called by
PROMPT_COMMAND causes bash not to display the PS1. Using backticks
instead of "$()" seems to work around the issue.

Here's an example of what fails (the prompt is never displayed):

    _setup_prompt() {
      foo=$(echo this can be anything)
      PS1="hello> "
    }
    PROMPT_COMMAND=_setup_prompt

But this works:

    _setup_prompt() {
      foo=`echo this can be anything`
      PS1="hello> "
    }
    PROMPT_COMMAND=_setup_prompt

WIP